+++ README.md
@@ -1,0 +1,5 @@
+Репозиторий для прохождения курса Otus Python Professional
+
+Выполненные проекты в рамках курса:
+- [Анализатор логов](src/log_analyzer/README.md)
+- To:do
