+++ configs/log_analyzer.yaml
@@ -1,4 +1,4 @@
 REPORT_SIZE: 100
-REPORT_DIR: homeworks/01_log_analyzer/reports
-LOG_DIR: data/01_log_analyzer
-LOG_FILE: homeworks/01_log_analyzer/logs/log.log
+REPORT_DIR: reports
+LOG_DIR: data
+LOG_FILE: logs/log.log
