+++ pyproject.toml
@@ -5,7 +5,7 @@ description = "Add your description here"
 readme = "README.md"
 requires-python = ">=3.13"
 dependencies = [
-        "pytest",
+    "pytest",
 ]
 
 [project.optional-dependencies]
@@ -14,7 +14,6 @@ dev = [
     "mypy",
     "ruff",
 ]
-
 
 [tool.ruff.lint]
 select = ["E4", "E7", "E9", "F"]
@@ -26,4 +25,10 @@ warn_return_any = true
 warn_unused_configs = true
 disallow_untyped_defs = true
 disallow_incomplete_defs = true
-disable_error_code = ["import-untyped"]
+disable_error_code = ["import-untyped", "import-not-found"]
+
+[tool.setuptools]
+package-dir = {"" = "src"}
+
+[tool.setuptools.packages.find]
+where = ["src"]
