--- tests/01_log_analyzer/test_01.py
@@ -1,5 +1,0 @@
-import time
-
-
-def test_1() -> None:
-    time.sleep(2)
